LPS-204580 Move `loadClientExtensions` to `frontend-js-web`
