+++ modules/apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/index.es.js
@@ -82,9 +82,11 @@ export {default as createPortletURL} from './liferay/util/portlet_url/create_por
 export {default as createRenderURL} from './liferay/util/portlet_url/create_render_url.es';
 export {default as createResourceURL} from './liferay/util/portlet_url/create_resource_url.es';
 
-// loadModule API
+// Client Extensions API
 
-export {getModuleAndSymbolNames, loadModule} from './module';
+export {getModuleAndSymbolNames} from './utils/client_extensions/getModuleAndSymbolNames';
+export {default as loadClientExtensions} from './utils/client_extensions/loadClientExtensions';
+export {loadModule} from './utils/client_extensions/loadModule';
 
 // Session API
