+++ modules/apps/frontend-js/frontend-js-web/types/src/main/resources/META-INF/resources/utils/client_extensions/getModuleAndSymbolNames.d.ts
@@ -1,0 +1,8 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+export declare function getModuleAndSymbolNames(
+	importDeclaration: string
+): [string, string];
