+++ modules/apps/frontend-js/frontend-js-web/types/src/main/resources/META-INF/resources/utils/client_extensions/loadClientExtensions.d.ts
