+++ modules/apps/frontend-js/frontend-js-web/types/src/main/resources/META-INF/resources/utils/client_extensions/loadModule.d.ts
@@ -3,9 +3,6 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-export declare function getModuleAndSymbolNames(
-	importDeclaration: string
-): [string, string];
 export declare function loadModule(
 	importDeclarationOrAMDModule: string
 ): Promise<any>;
